+++ packages/core/navigation-next/src/components/presentational/LayoutManager/types.js
@@ -32,6 +32,16 @@ export type ExperimentalFeatureFlags = {|
     mouse over the nav area whilst in a collapsed state.
   */
   experimental_flyoutOnHover: boolean,
+  /**
+    NOTE: This property is experimental and may be removed in a minor release.
+
+    FF to control the flyout activation area according to
+    https://product-fabric.atlassian.net/browse/NAV-197
+    This FF works in tandem with the other FF -> experimental_flyoutOnHover.
+    Setting this prop to true will have no effect unless the other is
+    set to true as well.
+  */
+  experimental_alternateFlyoutBehaviour: boolean,
 |};
 
 export type GetRefs = ({