+++ packages/editor/editor-core/src/__tests__/visual-regression/media/layouts.ts
@@ -12,6 +12,8 @@ export const mediaSingleLayouts = {
   'wrap-right': 'Wrap right',
   wide: 'Wide',
   'full-width': 'Full width',
+  'align-start': 'Align left',
+  'align-end': 'Align right',
 };
 
 describe('Snapshot Test: Media', () => {
@@ -42,7 +44,7 @@ describe('Snapshot Test: Media', () => {
         await page.waitForSelector(layoutButton);
         await page.click(layoutButton);
 
-        await page.waitForSelector(`.media-single.${layout}`);
+        await page.waitForSelector(`.media-single.image-${layout}`);
 
         await snapshot(page);
       }
@@ -73,7 +75,7 @@ describe('Snapshot Test: Media', () => {
         await page.waitForSelector(layoutButton);
         await page.click(layoutButton);
 
-        await page.waitForSelector(`.media-single.${layout}`);
+        await page.waitForSelector(`.media-single.image-${layout}`);
 
         await snapshot(page);
       }