+++ packages/editor/renderer/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/renderer
 
+## 32.2.0
+- [minor] [b9f8a8f](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b9f8a8f):
+
+  - Adding alignment options to media
+
 ## 32.1.2
 - [patch] [95f98cc](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/95f98cc):